backend/requirements.txt
fastapi
uvicorn
nltk
numpy
networkx
transformers
summarizer
tensorflow
pyjwt
pip